refactor(participant): inject common logic to participant abstract class
